biblioteca_db/src/controller/connection.py
import os
from dotenv import load_dotenv

class Settings:
    
    def __init__(self):
        self.db_host = os.getenv("DB_HOST")
        self.db_port = os.getenv("DB_PORT")
        self.db_user = os.getenv("DB_USER")
        self.db_pass = os.getenv("DB_PASS")
        self.db_name = os.getenv("DB_NAME")
        self.db_driver = os.getenv("DB_DRIVER")
        
    def database_url(self):
       return f"{self.db_driver}://{self.db_user}:{self.db_pass}@{self.db_host}:{self.db_port}/{self.db_name}"
    